Fix a bug in FileRefDatabase when deleting a non existent ref
Bug introduced in commit 0ae50001ef55d8cd44c715bdb62febe18c141739

+++ src/core/src/main/java/org/locationtech/geogig/storage/fs/FileRefDatabase.java
@@ -138,14 +138,16 @@ public class FileRefDatabase extends SimpleLockingRefDatabase {
         } catch (IllegalStateException symRefPointsToNonExistingRef) {
             oldRef = Optional.empty();
         }
-        Path refFile = toFile(refName);
-        try {
-            if (!Files.deleteIfExists(refFile)) {
-                throw new RuntimeException(
-                        "Unable to delete ref file '" + refFile.toAbsolutePath() + "'");
+        if (oldRef.isPresent()) {
+            Path refFile = toFile(refName);
+            try {
+                if (!Files.deleteIfExists(refFile)) {
+                    throw new RuntimeException(
+                            "Unable to delete ref file '" + refFile.toAbsolutePath() + "'");
+                }
+            } catch (IOException e) {
+                throw new IllegalStateException(e);
             }
-        } catch (IOException e) {
-            throw new IllegalStateException(e);
         }
         return RefChange.of(refName, oldRef, Optional.empty());
     }
